Add an :encode_full_message option to MessageVerifier to encode the full message
Previously only the payload is encoded. This is an issue if we want to easily
use a message in a url, or a non-case sensitive environment. We can change the
encoding, but this only changes the payload encoding. The digest (and separator)
remain unencoded.

This provides the option to fully encode a message, so the format changes from:

    encoded(serialized(message))--digest(serialized(message))

To:

    encoded(serialized(message)--digest(serialised(message)))